Make auto completion a little bit smarter
This will be changed soon, but for now when you open a file it gets all class, function and variable names and adds them to the completion list, but when you remove those variables in your code or close the tab, the list wont change. I will improve this feature in the near future.