Improved config.py
The use of config.py is more universal now and can be moved and edited with minimal interference from dev end.

Exceptions have been improved somehwat and it's fairly repetitive, but this is a good ground to get a standalone logging feature setup for PyPad and can be moved out once ready - the migration will be simple but will take time to make simple.

os.chdir was having problems on some environments grabbing cwd due to permissions. I've added a default path with this and only use that function in one exception when using config_reader(0) as it's the default configs. If it doesn't find them, elif blocks will head to either theme 1 or 2 by default. Most logic was already there, I just polished it.

Lmk if this is a shit push.